:sparkles: feat : 거리,시간 액티비티

+++ watch/app/src/main/java/com/example/gaenari/activity/dactivity/DFragmentStateAdapter.kt
@@ -4,7 +4,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 
-class DFragmentStateAdapter(fragmentActivity: FragmentActivity, private val programTarget: Int,private val programType: String,private val programTitle: String,private val programId: Long ) : FragmentStateAdapter(fragmentActivity) {
+class DFragmentStateAdapter(fragmentActivity: FragmentActivity, private val programTarget: Double,private val programType: String,private val programTitle: String,private val programId: Long ) : FragmentStateAdapter(fragmentActivity) {
     override fun getItemCount(): Int = 3
 
     override fun createFragment(position: Int): Fragment {
